Tests, refactoring, readme  - Create FloydWarshall test  - Create single route treatment test  - Create duplicate route treatment test  - Update CalculatedRouteList  - Create test for remote route set  - Create README.md  - Update pom to fix jackson security vulnerability
