+++ pom.xml
@@ -24,9 +24,10 @@
    <artifactId>network-simulator-api</artifactId>
    <version>1.0-SNAPSHOT</version>
 
-   <name>Patriots template project</name>
-   <description>This project is the default template, which will be used as starter for any other
-      project
+   <name>Patriot network simulator api</name>
+   <description>
+      This project is control api for network simulator project. Provides creating, connecting docker devices, networks,
+      calculating shortest paths and updating routers routing table.
    </description>
    <licenses>
       <license>
@@ -118,6 +119,8 @@
         <org.codehouse.mojo.version>1.6.0</org.codehouse.mojo.version>
         <org.args4j.version>2.33</org.args4j.version>
         <org.java-docker.version>3.1.0-rc-3</org.java-docker.version>
+        <com.fasterxml.jackson.core.version>2.8.11.1</com.fasterxml.jackson.core.version>
+        <com.github.jgonian.version>1.32</com.github.jgonian.version>
     </properties>
 
     <build>
@@ -159,7 +162,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <mainClass>com.redhat.patriot.network_controll.Main</mainClass>
+                    <mainClass>src/main/java/com/redhat/patriot/network/simulator/api/Main.java</mainClass>
                     <cleanupDaemonThreads>false</cleanupDaemonThreads>
                 </configuration>
             </plugin>
@@ -168,7 +171,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.redhat.patriot.network_controll.Main</mainClass>
+                            <mainClass>src/main/java/com/redhat/patriot/network/simulator/api/Main.java</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
@@ -230,12 +233,12 @@
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>2.6.4</version>
+            <version>${com.fasterxml.jackson.core.version}</version>
         </dependency>
         <dependency>
             <groupId>com.github.jgonian</groupId>
             <artifactId>commons-ip-math</artifactId>
-            <version>1.32</version>
+            <version>${com.github.jgonian.version}</version>
         </dependency>
     </dependencies>
 </project>
