+++ src/main/java/com/redhat/patriot/network/simulator/api/api/iptables/rules/Rule.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2018 Patriot project
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.redhat.patriot.network.simulator.api.api.iptables.rules;
+
+/**
+ * Ip tables rule.
+ */
+public interface Rule {
+    /**
+     * Create HTML path for ip tables rest request.
+     *
+     * @return the string
+     */
+    String toPath();
+}
