--- src/main/java/com/redhat/patriot/network_controll/Main.java
@@ -1,11 +1,0 @@
-package com.redhat.patriot.network_controll;
-
-
-public class Main {
-    public static void main(String[] args) {
-
-        FloydWarshallDemo floydWarshallDemo = new FloydWarshallDemo();
-        floydWarshallDemo.deploy();
-
-    }
-}
