--- src/main/java/com/redhat/patriot/network_controll/api/iptables/rules/Rule.java
@@ -1,13 +1,0 @@
-package com.redhat.patriot.network_controll.api.iptables.rules;
-
-/**
- * Ip tables rule.
- */
-public interface Rule {
-    /**
-     * Create HTML path for ip tables rest request.
-     *
-     * @return the string
-     */
-    String toPath();
-}
